Refactor MusicPlayer: Moved MusicControlButtons to a separate player_view.py file
